Renamed o-form-control-boolean.class file to o-boolean-form-data-component.class
