+++ projects/ontimize-web-ngx/src/lib/components/input/o-boolean-form-data-component.class.ts
